requirements.txt
streamlit
torch
numpy
opencv-python
Pillow
diffusers
transformers
accelerate
